pyproject.toml
[project]
name = "desktop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.4",
    "django-jet-reboot>=1.3.10",
    "django-unfold>=0.92.0",
    "djangorestframework>=3.17.1",
]
